+++ components/atoms/Marker/index.ts
@@ -1,0 +1,1 @@
+export { Marker } from './Marker';
