+++ components/molecules/MapSearch/MapSearch.styles.tsx
@@ -4,4 +4,5 @@ export const Container = styled.div`
   position: relative;
 
   width: 100%;
+  height: 100%;
 `;
